requirements.txt
fastapi>=0.95.0
sqlalchemy>=2.0.0
pydantic>=2.0
asyncpg>=0.21.0
python-dotenv>=1.0.0